project_requirements.txt
# AskiBill - Smart Expense Tracker Requirements
# Python 3.11+ required

# ===== Core Framework =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.47.1

# ===== Database & ORM =====
sqlalchemy==2.0.42
psycopg2-binary>=2.9.10
asyncpg==0.29.0
alembic==1.13.1

# ===== Authentication & Security =====
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.3.0
cryptography==41.0.8
python-multipart==0.0.6

# ===== Data Processing & Analytics =====
pandas>=2.3.1
numpy==1.25.2
plotly>=6.2.0
scikit-learn==1.3.2

# ===== HTTP Client & API Integration =====
httpx>=0.28.1
requests>=2.32.4
requests-oauthlib==1.3.1

# ===== Configuration & Environment =====
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== File Processing =====
pillow==10.1.0
aiofiles==23.2.1

# ===== Utilities =====
python-slugify==8.0.1
pytz==2023.3
python-dateutil==2.8.2
click==8.1.7

# ===== Development & Testing =====
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# ===== Optional: Banking Integration =====
# plaid-python==11.0.0  # Uncomment for Plaid integration
# stripe==7.8.0          # Uncomment for payment processing

# ===== Optional: Advanced Features =====
# redis==5.0.1           # Uncomment for session storage
# celery==5.3.4          # Uncomment for background tasks
# opencv-python==4.8.1.78  # Uncomment for receipt scanning
# pytesseract==0.3.10    # Uncomment for OCR functionality